+++ packages/commerce-sdk-react/src/hooks/ShopperCustomers/queryKeyHelpers.ts
@@ -6,7 +6,8 @@
  */
 import type {ShopperCustomers} from 'commerce-sdk-isomorphic'
 import {Argument, ExcludeTail} from '../types'
-import {pick} from '../utils'
+import {getCustomKeys, pick} from '../utils'
+import paramKeysMap from './paramKeys'
 
 // We must use a client with no parameters in order to have required/optional match the API spec
 type Client = ShopperCustomers<{shortCode: string}>
@@ -126,34 +127,24 @@ export type QueryKeys = {
 // This is defined here, rather than `types.ts`, because it relies on `Client` and `QueryKeys`,
 // and making those generic would add too much complexity.
 type QueryKeyHelper<T extends keyof QueryKeys> = {
-    /**
-     * Reduces the given parameters (which may have additional, unknown properties) to an object
-     * containing *only* the properties required for an endpoint.
-     */
-    parameters: (params: Params<T>) => Params<T>
     /** Generates the path component of the query key for an endpoint. */
     path: (params: Params<T>) => ExcludeTail<QueryKeys[T]>
     /** Generates the full query key for an endpoint. */
     queryKey: (params: Params<T>) => QueryKeys[T]
 }
 
-export const getExternalProfile: QueryKeyHelper<'getExternalProfile'> = {
-    parameters: (params) =>
-        pick(params, ['organizationId', 'externalId', 'authenticationProviderId', 'siteId']),
-    path: (params) => [
-        '/commerce-sdk-react',
-        '/organizations/',
-        params.organizationId,
-        '/customers/external-profile'
-    ],
-    queryKey: (params: Params<'getExternalProfile'>) => [
-        ...getExternalProfile.path(params),
-        getExternalProfile.parameters(params)
-    ]
-}
+// TODO: Re-implement (and update description from RAML spec) when the endpoint exits closed beta.
+// export const getExternalProfile: QueryKeyHelper<'getExternalProfile'> = {
+//     queryKey: (params: Params<'getExternalProfile'>) => {
+//         const paramKeys = [
+//             ...paramKeysMap['getExternalProfile'],
+//             ...getCustomKeys(getCustomerBaskets)
+//         ]
+//         return [...getExternalProfile.path(params), pick(params, paramKeys)]
+//     }
+// }
 
 export const getCustomer: QueryKeyHelper<'getCustomer'> = {
-    parameters: (params) => pick(params, ['organizationId', 'customerId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -161,14 +152,13 @@ export const getCustomer: QueryKeyHelper<'getCustomer'> = {
         '/customers/',
         params.customerId
     ],
-    queryKey: (params: Params<'getCustomer'>) => [
-        ...getCustomer.path(params),
-        getCustomer.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomer'>) => {
+        const paramKeys = [...paramKeysMap['getCustomer'], ...getCustomKeys(params)]
+        return [...getCustomer.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerAddress: QueryKeyHelper<'getCustomerAddress'> = {
-    parameters: (params) => pick(params, ['organizationId', 'customerId', 'addressName', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -178,14 +168,13 @@ export const getCustomerAddress: QueryKeyHelper<'getCustomerAddress'> = {
         '/addresses/',
         params.addressName
     ],
-    queryKey: (params: Params<'getCustomerAddress'>) => [
-        ...getCustomerAddress.path(params),
-        getCustomerAddress.parameters(params)
-    ]
-}
+    queryKey: (params: Params<'getCustomerAddress'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerAddress'], ...getCustomKeys(params)]
 
+        return [...getCustomerAddress.path(params), pick(params, paramKeys)]
+    }
+}
 export const getCustomerBaskets: QueryKeyHelper<'getCustomerBaskets'> = {
-    parameters: (params) => pick(params, ['organizationId', 'customerId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -194,25 +183,13 @@ export const getCustomerBaskets: QueryKeyHelper<'getCustomerBaskets'> = {
         params.customerId,
         '/baskets'
     ],
-    queryKey: (params: Params<'getCustomerBaskets'>) => [
-        ...getCustomerBaskets.path(params),
-        getCustomerBaskets.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerBaskets'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerBaskets'], ...getCustomKeys(params)]
+        return [...getCustomerBaskets.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerOrders: QueryKeyHelper<'getCustomerOrders'> = {
-    parameters: (params) =>
-        pick(params, [
-            'organizationId',
-            'customerId',
-            'crossSites',
-            'from',
-            'until',
-            'status',
-            'siteId',
-            'offset',
-            'limit'
-        ]),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -221,15 +198,13 @@ export const getCustomerOrders: QueryKeyHelper<'getCustomerOrders'> = {
         params.customerId,
         '/orders'
     ],
-    queryKey: (params: Params<'getCustomerOrders'>) => [
-        ...getCustomerOrders.path(params),
-        getCustomerOrders.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerOrders'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerOrders'], ...getCustomKeys(params)]
+        return [...getCustomerOrders.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerPaymentInstrument: QueryKeyHelper<'getCustomerPaymentInstrument'> = {
-    parameters: (params) =>
-        pick(params, ['organizationId', 'customerId', 'paymentInstrumentId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -239,14 +214,16 @@ export const getCustomerPaymentInstrument: QueryKeyHelper<'getCustomerPaymentIns
         '/payment-instruments/',
         params.paymentInstrumentId
     ],
-    queryKey: (params: Params<'getCustomerPaymentInstrument'>) => [
-        ...getCustomerPaymentInstrument.path(params),
-        getCustomerPaymentInstrument.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerPaymentInstrument'>) => {
+        const paramKeys = [
+            ...paramKeysMap['getCustomerPaymentInstrument'],
+            ...getCustomKeys(params)
+        ]
+        return [...getCustomerPaymentInstrument.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerProductLists: QueryKeyHelper<'getCustomerProductLists'> = {
-    parameters: (params) => pick(params, ['organizationId', 'customerId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -255,14 +232,13 @@ export const getCustomerProductLists: QueryKeyHelper<'getCustomerProductLists'> 
         params.customerId,
         '/product-lists'
     ],
-    queryKey: (params: Params<'getCustomerProductLists'>) => [
-        ...getCustomerProductLists.path(params),
-        getCustomerProductLists.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerProductLists'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerProductLists'], ...getCustomKeys(params)]
+        return [...getCustomerProductLists.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerProductList: QueryKeyHelper<'getCustomerProductList'> = {
-    parameters: (params) => pick(params, ['organizationId', 'customerId', 'listId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -272,15 +248,13 @@ export const getCustomerProductList: QueryKeyHelper<'getCustomerProductList'> = 
         '/product-lists/',
         params.listId
     ],
-    queryKey: (params: Params<'getCustomerProductList'>) => [
-        ...getCustomerProductList.path(params),
-        getCustomerProductList.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerProductList'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerProductList'], ...getCustomKeys(params)]
+        return [...getCustomerProductList.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getCustomerProductListItem: QueryKeyHelper<'getCustomerProductListItem'> = {
-    parameters: (params) =>
-        pick(params, ['organizationId', 'customerId', 'listId', 'itemId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -292,30 +266,30 @@ export const getCustomerProductListItem: QueryKeyHelper<'getCustomerProductListI
         '/items/',
         params.itemId
     ],
-    queryKey: (params: Params<'getCustomerProductListItem'>) => [
-        ...getCustomerProductListItem.path(params),
-        getCustomerProductListItem.parameters(params)
-    ]
+    queryKey: (params: Params<'getCustomerProductListItem'>) => {
+        const paramKeys = [...paramKeysMap['getCustomerProductListItem'], ...getCustomKeys(params)]
+        return [...getCustomerProductListItem.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getPublicProductListsBySearchTerm: QueryKeyHelper<'getPublicProductListsBySearchTerm'> =
     {
-        parameters: (params) =>
-            pick(params, ['organizationId', 'email', 'firstName', 'lastName', 'siteId']),
         path: (params) => [
             '/commerce-sdk-react',
             '/organizations/',
             params.organizationId,
             '/product-lists'
         ],
-        queryKey: (params: Params<'getPublicProductListsBySearchTerm'>) => [
-            ...getPublicProductListsBySearchTerm.path(params),
-            getPublicProductListsBySearchTerm.parameters(params)
-        ]
+        queryKey: (params: Params<'getPublicProductListsBySearchTerm'>) => {
+            const paramKeys = [
+                ...paramKeysMap['getPublicProductListsBySearchTerm'],
+                ...getCustomKeys(params)
+            ]
+            return [...getPublicProductListsBySearchTerm.path(params), pick(params, paramKeys)]
+        }
     }
 
 export const getPublicProductList: QueryKeyHelper<'getPublicProductList'> = {
-    parameters: (params) => pick(params, ['organizationId', 'listId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -323,14 +297,13 @@ export const getPublicProductList: QueryKeyHelper<'getPublicProductList'> = {
         '/product-lists/',
         params.listId
     ],
-    queryKey: (params: Params<'getPublicProductList'>) => [
-        ...getPublicProductList.path(params),
-        getPublicProductList.parameters(params)
-    ]
+    queryKey: (params: Params<'getPublicProductList'>) => {
+        const paramKeys = [...paramKeysMap['getPublicProductList'], ...getCustomKeys(params)]
+        return [...getPublicProductList.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getProductListItem: QueryKeyHelper<'getProductListItem'> = {
-    parameters: (params) => pick(params, ['organizationId', 'listId', 'itemId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
@@ -340,8 +313,8 @@ export const getProductListItem: QueryKeyHelper<'getProductListItem'> = {
         '/items/',
         params.itemId
     ],
-    queryKey: (params: Params<'getProductListItem'>) => [
-        ...getProductListItem.path(params),
-        getProductListItem.parameters(params)
-    ]
+    queryKey: (params: Params<'getProductListItem'>) => {
+        const paramKeys = [...paramKeysMap['getProductListItem'], ...getCustomKeys(params)]
+        return [...getProductListItem.path(params), pick(params, paramKeys)]
+    }
 }