+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## v2.4.0-dev (Jan 22, 2024)
 
+### Bug Fixes
+
+- Fix invalid query params warnings [#1655](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1655)
+- Fix `product-item` component imports to ensure that it is overridable. [#1672](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1672)
+- Fix locale selector navigating back to default locale [#1670](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1670)
+
 ## v2.3.1 (Jan 23, 2024)
 
 ### Bug Fixes